move offers table to common ui
